pyproject.toml
[project]
name = "etl"
version = "0.1.0"
description = "CSV ETL CLI Tool"
authors = [
    { name = "Manideep" }
]
dependencies = [
    "pandas",
    "pyarrow",
    "pyyaml"
]
requires-python = ">=3.8"


[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "ruff"
]

[project.scripts]
etl = "etl.cli:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]